+++ src/index.ts
@@ -55,6 +55,7 @@ export interface ReleasePROptions {
   // and then remove this property
   versionFile?: string;
   pullRequestTitlePattern?: string;
+  signoff?: string;
   extraFiles?: string[];
 }
 
@@ -72,6 +73,7 @@ interface ReleaserConstructorOptions {
 interface ManifestOptions {
   configFile?: string;
   manifestFile?: string;
+  signoff?: string;
 }
 
 export interface ManifestConstructorOptions